Some progress - managed to eliminate the infinite update loops, although now I've partially broken [again] the delay behaviour.
I think to get the right behaviour I need to keep a record of sizes the view renders at; when it last rendered for each size.  Ordered descendingly (somehow… not sure how to handle two dimensions simultaneously).  Then it can shrink past each size only after the delay; it'll pause as necessary at each size interval in order to enforce the desired delay (as opposed to how it currently behaves which is to just immediately shrink if it's already actively shrinking and the desired size shrinks).  But that would seemingly require me to distinguish shrink updates (re-layout and re-draw) from externally-triggered ones, as I don't want to remember the drawn size of the shrinks… 🤔

I've managed to simplify things a lot, removing all but the 'tick' from the wrapper View (which seems to be the bare minimum).  There's a lot of stuff still in sizeThatFits which is probably unnecessary - I haven't gone back and pruned it yet.