Fixed how `currentMinimumSize` works, to accomodate height changes too.
Previously it was ignoring the height, and just picking the widest still-alive rect, as its result.  But that rect isn't necessarily the tallest.  Now, it just returns the union of every alive rect, which is a lot more conservative but yields the expected results.

This might hurt the performance, alas, because it can no longer early-exit as soon as it finds the biggest frame.  Then again, it no longer sorts `renderTimesPerDesiredSize` into a new `Array` every time it's called, so possibly it's actually faster now than it used to be, at least in some cases. 🤷‍♂️

Though it does now return a much more conservative (and often inaccurately low) minimum delay, too - it just returns the duration until the next entry expires.  I assume it's possible in principle to optimise this, at least algorithmically, but I'm not sure if it'd *actually* improve performance because the 'optimisation' would have to consider the cached rect's relative sizes and overlaps, which might be more trouble than it's worth.

For now, in any case, the performance seems - before & after this change - "good enough"; even my shitty old iMac Pro can handle a dozen Viagra'd views animating at 60 Hz.  And that's in a debug build, too.